requirements.txt
jsonschema==3.2.0
numpy==1.19.2
pandas==1.2.0
scipy==1.5.2
torch==1.7.1
seaborn==0.11.1
matplotlib==3.3.2
coloredlogs==15.0
progressbar2==3.53.1
pytest==6.2.2
scikit_learn==0.23.2
sklearn_evaluation==0.5.4
ipython==7.19.0
black==20.8b1
